Add files via upload
First Upload of Prony Material. Attempts to do some analysis of noise. Has a calculator that allows for computation of Prony; allows for case based testing. In Prony Test Suite; there is a noise incrementer, where you can see the norm of the estimated versus actual mode frequencies increase. More work is needed in order to correctly deal with a measure of the error for a non equal number of guessed and estimated exponential. More experimentation should be done with moving the frequencies of different modes closer together and the relationship between how close the frequencies of two modes are together, various noise levels, resolution of damping factors for the modes and amplitude of the modes. Believe this can be accomplished empirically by manually moving the mode frequencies closer together and looking at results, documenting test cases.