Add files via upload
Five Minute Mess around with n4sid. Will continue to work with it and read documentation tomorrow morning. Have done more reading on subspace methods. Need to find out more about regularization is done in this framework. If it both done in final estimation of A, B, C , D and/or if it is done in the estimation of THETA in order to get n k-step ahead predictors